+++ internal/services/user_service.go
@@ -15,11 +15,11 @@ func NewUserService(repo repositories.UserRepository) *UserService {
 	return &UserService{userRepo: repo}
 }
 
-func (u *UserService) GetUser(ctx context.Context, id int) (*models.User, error) {
+func (u *UserService) GetUser(ctx context.Context, id int64) (*models.User, error) {
 	return u.userRepo.GetUserByID(ctx, id)
 }
 
-func (u *UserService) GetUserRolesByID(ctx context.Context, id int) (*models.UserRoles, error) {
+func (u *UserService) GetUserRolesByID(ctx context.Context, id int64) (*models.UserRoles, error) {
 
 	roles, err := u.userRepo.GetUserRolesByID(ctx, id)
 	if roles != nil {
